# docs/swagger.yaml
definitions:
  handlers.CompleteUploadInput:
    properties:
      files:
        items:
          properties:
            contentType:
              type: string
            filename:
              type: string
            key:
              type: string
            size:
              type: integer
          type: object
        type: array
      token:
        type: string
    type: object
  handlers.PresignResponse:
    properties:
      token:
        type: string
      urls:
        items:
          $ref: '#/definitions/handlers.PresignedFile'
        type: array
    type: object
  handlers.PresignedFile:
    properties:
      filename:
        type: string
      key:
        type: string
      uploadURL:
        type: string
    type: object
  utils.Payload:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/files/complete:
    post:
      consumes:
      - application/json
      description: Verifies uploaded files on R2, stores file metadata, and registers
        the upload session in the database. Each transfer is valid for 1 hour and
        limited to 100MB for anonymous uploads.
      parameters:
      - description: Upload completion payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CompleteUploadInput'
      produces:
      - application/json
      responses:
        "200":
          description: Files uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Payload'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Invalid input or verification failed
          schema:
            $ref: '#/definitions/utils.Payload'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Complete file upload
      tags:
      - Files
  /api/v1/files/presign:
    post:
      consumes:
      - application/json
      description: Accepts a list of files (name and size), validates the total size,
        and returns presigned PUT URLs for each file. Each upload session is identified
        by a unique token.
      parameters:
      - description: List of files to upload
        in: body
        name: input
        required: true
        schema:
          items:
            properties:
              filename:
                type: string
              size:
                type: integer
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Presigned URLs generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Payload'
            - properties:
                data:
                  $ref: '#/definitions/handlers.PresignResponse'
              type: object
        "400":
          description: Invalid input or size limit exceeded
          schema:
            $ref: '#/definitions/utils.Payload'
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Failed to generate presigned URL
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Generate presigned URLs for file upload
      tags:
      - Files
  /api/v1/share/{token}:
    get:
      consumes:
      - application/json
      description: Returns metadata (name, size, contentType, index) of all files
        in a shared transfer.
      parameters:
      - description: Share token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Files retrieved successfully
          schema:
            $ref: '#/definitions/utils.Payload'
        "400":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Invalid or expired share link
          schema:
            $ref: '#/definitions/utils.Payload'
        "410":
          description: Share link has expired
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Retrieve shared file details
      tags:
      - Share
  /api/v1/share/{token}/presign-download/{index}:
    get:
      consumes:
      - application/json
      description: Returns a temporary signed URL to download a specific file (by
        index) from a shared transfer.
      parameters:
      - description: Share token
        in: path
        name: token
        required: true
        type: string
      - description: File index
        in: path
        name: index
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Presigned download URL generated successfully
          schema:
            $ref: '#/definitions/utils.Payload'
        "400":
          description: Missing or invalid parameters
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: File not found or invalid share link
          schema:
            $ref: '#/definitions/utils.Payload'
        "410":
          description: Share link has expired
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Generate a presigned download URL
      tags:
      - Share
swagger: "2.0"